Optimization on entry.sh
Cloning a full repository in order to retrieve a log of it is unnecessary. Logging remotely is not supported so instead retrieve the last 3 commits in the clone process. This will cut down the run time of the solution massively (especially when scanning large repositories) because much less download is required.